link up model with interface
