+++ app/scripts/interface.js
@@ -1,25 +1,29 @@
 'use strict';
 
 $(document).ready(function() {
+  var template = $('#task-items-container').html();
   var taskManager = new TaskManager();
-  $('#task-form').on('submit', function(event) {
+
+
+  runRenderView();
+
+  $('body').on('submit', '#task-form', function(event) {
     event.preventDefault();
     var text = $('#task-content').val();
+    $('#task-content').val('');
     taskManager.addTask(text);
+    runRenderView();
+  })
 
-    $('#pending-task-items').empty();
-    for(var i = 0; i < taskManager.pendingTasks.length; i++) {
-      $('#pending-task-items').append('<li>' + '<input type=\'checkbox\'' + ' ' + 'name=' + i + ' ' + 'id=' + i +'>' + taskManager.pendingTasks[i] + '</li>');
-    }
-
-      $(':checkbox').change(function() {
-        for(var i = 0; i < taskManager.pendingTasks.length; i++) {
-          if ($('#' + i).is(":checked")) {
-          taskManager.markAsComplete(taskManager.pendingTasks[i]);
-          $('#completed-task-items').append('<li>' + taskManager.completedTasks[i] + '</li>');
-        }
-      }
-    });
+  $('body').on('change', ':checkbox', function(event) {
+    var eventID = event.currentTarget.name;
+    taskManager.markAsComplete(taskManager.pendingTasks[eventID]);
+    runRenderView();
   });
 
+  function runRenderView() {
+    renderView(taskManager, template, function(renderedContent) {
+      $('#task-items-container').html(renderedContent);
+    });
+  };
 });
